% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_95ci.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratios and 95% Confidence Intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{A vector of coefficient estimates.}

\item{se}{A vector of standard errors.}

\item{siglevel}{The significance level (0.05 for 95% confidence).}

\item{roundto}{The number of decimal places to round to.}
}
\value{
A string that contains odds ratios and 95% confidence intervals.
}
\description{
A function that calculates odds ratios and 95% confidence intervals
from a vector of coefficient estimates and a vector of standard errors.
}
\examples{
OR_95CI(coef = c(0.5, -0.3), se = c(0.1, 0.15), siglevel = 0.05, roundto = 2)
}
\author{
Anna He
}
